liberdus_proxy::liberdus
Function handle_request

pub async fn handle_request<S>(
    request_buffer: Vec<u8>,
    client_stream: &mut S,
    liberdus: Arc<Liberdus>,
    config: Arc<Config>,
) -> Result<(), Box<dyn Error>>
where
    S: AsyncWrite + AsyncRead + Unpin + Send,

Handles the client request stream by reading the request, forwarding it to a consensor server, collect the response from validator, and relaying it back to the client. Handles a single HTTP request from the client stream. This function assumes it processes only one http payload request. The proxy will maintain the tcp stream open for a specify amount of time to avoid the handshake overhead when the client do frequent calls. Proxy will not maintain tcp stream connected to upstream server/validator. It is always single use and always shutdown after the response is sent back to the client. If the multiple request from single client tcp stream, it is advisable to pick different validator each time.

No chunked encoding is supported. No Multiplexing is supported.