Module liberdus_proxy::http

Client Request Handler Module

This module defines the core logic for handling client requests in a TCP-based system. It handles the communication between clients and backend servers (consensors) via a load-balancing system. The module uses asynchronous I/O operations provided by Tokio, ensuring scalability and efficiency.

Features

Reads and parses HTTP requests from clients with support for timeouts.

Forwards client requests to a backend consensor server.

Relays server responses back to the client.

Implements error handling for timeouts, unavailable servers, and unexpected I/O errors.

Components

handle_stream: The main entry point for processing client requests.

read_or_collect: Reads and collects request or response data with header parsing.

respond_with_internal_error: Sends a 500 Internal Server Error response to the client.

respond_with_timeout: Sends a 504 Gateway Timeout response to the client.

Functions

handle_request

Handles the client request stream by reading the request, forwarding it to a consensor server, collect the response from validator, and relaying it back to the client. Handles a single HTTP request from the client stream. This function assumes it processes only one http payload request. The proxy will maintain the tcp stream open for a specify amount of time to avoid the handshake overhead when the client do frequent calls. Proxy will not maintain tcp stream connected to upstream server/validator. It is always single use and always shutdown after the response is sent back to the client. If the multiple request from single client tcp stream, it is advisable to pick different validator each time.

handle_stream

Outer loop to handle multiple HTTP requests from the same stream

parse_content_length 🔒

Parses the Content-Length header from the given headers. Returns None if the header is missing or invalid.

read_or_collect

Reads from the stream until the end of the headers or the end of the body if the Content-Length header is present. The data is collected into the buffer.

respond_with_internal_error 🔒

Takes the stream, responds with a 500 Internal Server Error, and shutdown tcp

respond_with_timeout 🔒

Takes the stream, responds with a timeout error, and shutdown tcp

set_http_header

Helper function to insert or replace a header in the HTTP response buffer.