liberdus_proxy::shardus_monitor
Module proxy

Structs

PayloadCache

Statics

CACHE_MAP

Functions

handle_request

Handles the client request stream by reading the request, forwarding it to a consensor server, collect the response from validator, and relaying it back to the client. Handles a single HTTP request from the client stream. This function assumes it processes only one http payload request. The proxy will maintain the tcp stream open for a specify amount of time to avoid the handshake overhead when the client do frequent calls. Proxy will not maintain tcp stream connected to upstream server/validator. It is always single use and always shutdown after the response is sent back to the client. If the multiple request from single client tcp stream, it is advisable to pick different validator each time.

is_monitor_route